기능: 소셜 로그인 서비스 작업

+++ api/src/social-user/social-user.service.ts
@@ -1,4 +1,5 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { SocialUser } from '@prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateSocialUserDto } from './dto/create-social-user.dto';
 import { DeleteSocialUserDto } from './dto/delete-social-user.dto';
@@ -13,10 +14,7 @@ export class SocialUserService {
    * @param providerAccountId - 소셜 유저 provider 계정 ID
    * @returns 소셜 유저 정보
    */
-  async findByProvider(
-    provider: string,
-    providerAccountId: string,
-  ): Promise<any> {
+  async findByProvider(provider: string, providerAccountId: string): Promise<SocialUser> {
     const user = await this.prisma.socialUser.findUnique({ where: { provider, providerAccountId } });
     if (!user) throw new NotFoundException('소셜 유저를 찾을 수 없습니다.');
     return user;
@@ -28,10 +26,10 @@ export class SocialUserService {
    * @returns 소셜 유저 정보
    * @throws BadRequestException - 소셜 유저가 이미 존재하는 경우
    */
-  async createSocialUser(dto: CreateSocialUserDto): Promise<any> {
+  async createSocialUser(dto: CreateSocialUserDto): Promise<SocialUser> {
     const { userIdx, provider, providerAccountId } = dto;
 
-    const exists = await this.findByProvider(provider, providerAccountId);
+    const exists = await this.prisma.socialUser.findUnique({ where: { provider, providerAccountId } });
     if (exists) throw new BadRequestException('이미 존재하는 소셜 유저입니다.');
 
     const socialUser = await this.prisma.socialUser.create({
@@ -41,9 +39,8 @@ export class SocialUserService {
         providerAccountId
       },
     });
-  
     return socialUser;
-  };
+  }
 
   /**
    * 소셜 유저를 삭제합니다.
@@ -51,16 +48,13 @@ export class SocialUserService {
    * @returns 삭제된 소셜 유저 정보
    * @throws NotFoundException - 소셜 유저를 찾을 수 없는 경우
    */
-  async deleteSocialUser(dto: DeleteSocialUserDto): Promise<any> {
-    const { provider, providerAccountId } = dto;
+  async deleteSocialUser(dto: DeleteSocialUserDto): Promise<SocialUser> {
+    const { userIdx } = dto;
 
-    const socialUser = await this.findByProvider(provider, providerAccountId);
+    const socialUser = await this.prisma.socialUser.findUnique({ where: { userIdx } });
     if (!socialUser) throw new NotFoundException('소셜 유저를 찾을 수 없습니다.');
 
-    const deletedSocialUser = await this.prisma.socialUser.delete({
-      where: { provider, providerAccountId },
-    });
-
+    const deletedSocialUser = await this.prisma.socialUser.delete({ where: { userIdx } });
     return deletedSocialUser;
   }
 }
